backend/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.8.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# MCP and LLM
mcp>=1.0.0
anthropic>=0.18.0

# CORS and middleware
python-multipart==0.0.6

# Authentication and Database
authlib>=1.3.0
itsdangerous>=2.1.2
python-jose[cryptography]>=3.3.0
aiomysql>=0.2.0
sqlalchemy>=2.0.0
boto3>=1.34.0
cryptography>=41.0.0
PyMySQL>=1.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx>=0.27.0

# Development
black==23.12.1
ruff==0.1.11
mypy==1.8.0
